# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carteira Digital API",
        "description": "API para gerenciamento de carteira financeira (registro, login, usuário, logout).",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um usuário",
                "description": "Registra um novo usuário e cria sua carteira zerada.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário registrado com sucesso"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "UserA"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "a@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login do usuário",
                "description": "Autentica o usuário e retorna um token Bearer.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retorna o usuário autenticado",
                "description": "Retorna os dados do usuário autenticado (incluindo wallet), identificando-o pelo token.",
                "operationId": "7c1cf647d76fad4b3eb1549fa4fe2f4c",
                "responses": {
                    "200": {
                        "description": "Dados do usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout do usuário",
                "description": "Revoga todos os tokens do usuário (logout).",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout efetuado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout efetuado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deposit": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Depósito na carteira",
                "description": "Executa um depósito.",
                "operationId": "7d37929cec1ed6739970c3ba17c8102b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Depósito realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Depósito realizado com sucesso"
                                        },
                                        "transactionId": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "transaction": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou saldo inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transfer": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Transferência de saldo",
                "description": "Executa uma transferência.",
                "operationId": "ff447720cf8647ed4d7a3854209f0b69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transferência realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transferência realizada com sucesso"
                                        },
                                        "transactionId": {
                                            "type": "integer",
                                            "example": 11
                                        },
                                        "transaction": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Saldo insuficiente ou erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transações"
                ],
                "summary": "Histórico de transações",
                "description": "Lista o histórico de transações.",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "responses": {
                    "200": {
                        "description": "Lista de transações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransactionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reverse/{transaction}": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Reversão de transação",
                "description": "Reverte uma transação via route model binding.",
                "operationId": "57ce17061ab13d90fc383130833abae0",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "ID da transação a ser revertida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transação revertida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transação revertida com sucesso"
                                        },
                                        "transactionId": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "reversal": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Operação não permitida"
                    },
                    "422": {
                        "description": "Transação já revertida ou erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DepositRequest": {
                "required": [
                    "amount"
                ],
                "properties": {
                    "amount": {
                        "description": "Valor a ser depositado, maior que zero",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Nome completo do usuário",
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "description": "Email válido do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "password": {
                        "description": "Senha com no mínimo 8 caracteres",
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
            "TransferRequest": {
                "required": [
                    "payee_id",
                    "amount"
                ],
                "properties": {
                    "payee_id": {
                        "description": "ID do usuário destinatário",
                        "type": "integer",
                        "example": 2
                    },
                    "amount": {
                        "description": "Valor a ser transferido, maior que zero",
                        "type": "number",
                        "format": "float",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "TransactionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "type": {
                        "type": "string",
                        "example": "deposit"
                    },
                    "amount": {
                        "type": "string",
                        "example": "100,00"
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "payer_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "payee_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "wallet": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "balance": {
                                "type": "string",
                                "example": "100,00"
                            },
                            "updated": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-04-23 15:00:00"
                            }
                        },
                        "type": "object"
                    },
                    "joined": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23 15:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Transações",
            "description": "Operações de depósito, transferência, histórico e reversão"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}